refactor(MonthlyScrapDeadlineResponse) : 중첩 DTO를 동일 파일 내 평면 구조로 분리
