--- src/main/kotlin/com/terning/server/kotlin/application/scrap/dto/MonthlyScrapDeadLineSummary.kt
@@ -1,7 +1,0 @@
-package com.terning.server.kotlin.application.scrap.dto
-
-data class MonthlyScrapDeadLineSummary(
-    val scrapId: Long,
-    val title: String,
-    val color: String,
-)
