--- src/main/kotlin/com/terning/server/kotlin/application/scrap/dto/MonthlyScrapDeadlineGroup.kt
@@ -1,6 +1,0 @@
-package com.terning.server.kotlin.application.scrap.dto
-
-data class MonthlyScrapDeadlineGroup(
-    val deadline: String,
-    val scraps: List<MonthlyScrapDeadLineSummary>,
-)
